requirements.txt
# IMDB Sentiment Analysis - Requirements
# Core dependencies for GPT-2 based sentiment classification

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers
transformers>=4.30.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning Utilities
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bar
tqdm>=4.65.0

# Configuration (optional)
pyyaml>=6.0
